chore: restructure backend services and add React UI
