+++ data-agent/Dockerfile.api
@@ -1,0 +1,6 @@
+FROM python:3.11-slim
+WORKDIR /app
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+COPY app ./app
+CMD ["uvicorn", "app.api:app", "--host", "0.0.0.0", "--port", "8000"]
