+++ docker-compose.yml
@@ -1,0 +1,23 @@
+version: '3.9'
+services:
+  api:
+    build:
+      context: ./data-agent
+      dockerfile: Dockerfile.api
+    volumes:
+      - ./data-agent/app:/app/app
+    ports:
+      - "8000:8000"
+  frontend:
+    build: ./frontend
+    ports:
+      - "3000:80"
+  ollama:
+    image: ollama/ollama:latest
+    restart: unless-stopped
+    ports:
+      - "11434:11434"
+    volumes:
+      - ollama_data:/root/.ollama
+volumes:
+  ollama_data:
