--- frontend/src/App.jsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import { Container, Typography } from '@mui/material';
-import { BarChart, Bar, XAxis, YAxis, Tooltip, ResponsiveContainer } from 'recharts';
-
-const data = [
-  { name: 'A', value: 12 },
-  { name: 'B', value: 18 },
-  { name: 'C', value: 5 }
-];
-
-export default function App() {
-  return (
-    <Container sx={{ marginTop: 4 }}>
-      <Typography variant="h4" gutterBottom>
-        Data Agent React UI
-      </Typography>
-      <ResponsiveContainer width="100%" height={300}>
-        <BarChart data={data}>
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Tooltip />
-          <Bar dataKey="value" fill="#1976d2" />
-        </BarChart>
-      </ResponsiveContainer>
-    </Container>
-  );
-}
